variables for pin definitions changed from int8_t to int16_t to make this comatible with stm32duino (tested using 2.1.0)

+++ Adafruit_ST77xx.h
@@ -88,13 +88,13 @@
 /// Subclass of SPITFT for ST77xx displays (lots in common!)
 class Adafruit_ST77xx : public Adafruit_SPITFT {
 public:
-  Adafruit_ST77xx(uint16_t w, uint16_t h, int8_t _CS, int8_t _DC, int8_t _MOSI,
-                  int8_t _SCLK, int8_t _RST = -1, int8_t _MISO = -1);
-  Adafruit_ST77xx(uint16_t w, uint16_t h, int8_t CS, int8_t RS,
-                  int8_t RST = -1);
+  Adafruit_ST77xx(uint16_t w, uint16_t h, int16_t _CS, int16_t _DC, int16_t _MOSI,
+                  int16_t _SCLK, int16_t _RST = -1, int16_t _MISO = -1);
+  Adafruit_ST77xx(uint16_t w, uint16_t h, int16_t CS, int16_t RS,
+                  int16_t RST = -1);
 #if !defined(ESP8266)
-  Adafruit_ST77xx(uint16_t w, uint16_t h, SPIClass *spiClass, int8_t CS,
-                  int8_t RS, int8_t RST = -1);
+  Adafruit_ST77xx(uint16_t w, uint16_t h, SPIClass *spiClass, int16_t CS,
+                  int16_t RS, int16_t RST = -1);
 #endif // end !ESP8266
 
   void setAddrWindow(uint16_t x, uint16_t y, uint16_t w, uint16_t h);
